Fix mock of Client always expecting self::SOME_YEAR
With the old way, using a test class property as mock instace
for the client, created in the setUp(), it wasn't possible
to expect an other year than SOME_YEAR.
Even if you didn't inject the property into the SeasonService
and instead created a fresh mock in a local var and injected
that local var, it still seemed to use the mock from the setUp
method.

This doesn't make any sense but at lease using a builder method
instead fixes this.